fix: rating modal overriding tab pages
