+++ apps/mobile/src/app/(tabs)/(profile)/_layout.tsx
@@ -11,32 +11,7 @@ const Layout = () => {
 				animation: "fade",
 				title: "",
 			}}
-		>
-			<Stack.Screen
-				name="(modals)/rating"
-				options={{
-					title: "",
-					presentation: "modal",
-					animation: "slide_from_bottom",
-				}}
-			/>
-			<Stack.Screen
-				name="(modals)/reply/rating"
-				options={{
-					title: "",
-					presentation: "modal",
-					animation: "slide_from_bottom",
-				}}
-			/>
-			<Stack.Screen
-				name="(modals)/reply/comment"
-				options={{
-					title: "",
-					presentation: "modal",
-					animation: "slide_from_bottom",
-				}}
-			/>
-		</Stack>
+		/>
 	);
 };
